+++ loginBackend/src/main/java/com/nus/iss/login/repository/UserProfileRepository.java
@@ -4,6 +4,6 @@ import com.nus.iss.login.entity.UserProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserProfileRepository extends JpaRepository<UserProfile, Long> {
-    Optional<UserProfile> findByUserUserID(Long userId);
+public interface UserProfileRepository extends JpaRepository<UserProfile, Integer> {
+    Optional<UserProfile> findByUserUserID(Integer userId);
 }
